+++ src/utils/index.js
@@ -1,12 +1,12 @@
-const { accessSync } = require('fs')
+import { accessSync } from 'fs'
 
-const MIN_CHROME_VERSION = 85
+export const MIN_CHROME_VERSION = 85
 
-const MAX_CHROME_VERSION = Number.MAX_VALUE
+export const MAX_CHROME_VERSION = Number.MAX_VALUE
 
-const newLineRegex = /\r?\n/
+export const newLineRegex = /\r?\n/
 
-function canAccess(file) {
+export function canAccess(file) {
   if (!file) {
     return false
   }
@@ -17,11 +17,4 @@ function canAccess(file) {
   } catch (e) {
     return false
   }
-}
-
-module.exports = {
-  MIN_CHROME_VERSION,
-  MAX_CHROME_VERSION,
-  newLineRegex,
-  canAccess
 }